iln/python_src/requirements.txt
einops==0.4.1
matplotlib==3.5.1
numpy==1.22.3
PyYAML==6.0
tqdm==4.63.0